Squish duplicates when running spearman
If we don't squish the duplicates, then the duplicates mess up our
correlation and cause even identical files to be only ~85% correlated.

This commit also breaks out `assertMostlyAgrees` into a separate file so
that we do not have to repeat the logic in the tests for every different
processor, and it breaks out `squish_duplicates` into a function in
`utils.py` so that we can use the same squishing method in both the
smasher and the tests.
